+++ src/functional/constant.ts
@@ -1,7 +1,6 @@
 /**
- * A function that returns a function, that returns `x`
+ * Returns `() => x`
  */
-
 export const constant =
 	<T = any>(x: T) =>
 	() =>