+++ src/functional/functional.ts
@@ -1,11 +1,12 @@
-import { last, lastOut, substitute, Tuple } from "../array/array.js"
-import { not, T } from "../boolean/boolean.js"
+import { last, lastOut, substitute } from "../array/array.js"
+import { not } from "../boolean/boolean.js"
 import { max } from "../number/number.js"
 import {
 	AnyFunction,
 	ArrayMapper,
 	BasicFunction,
 	Falsy,
+	Filter,
 	isUndefined,
 	MapReturnType,
 } from "../types/types.js"
@@ -13,6 +14,7 @@ import {
 /**
  * Provided with a list of functions, lazily executes them in sequence
  * and returns the first truthy result.
+ *
  * If all results are falsy - returns the last one
  */
 export const or =
@@ -23,6 +25,7 @@ export const or =
 /**
  * Provided with a list of functions, lazily executes them in sequence
  * and returns the first falsy result.
+ *
  * If all results are truthy - returns the last one
  */
 export const and =
@@ -31,46 +34,50 @@ export const and =
 		fs.reduce((prev, curr) => (prev ? curr(...x) : prev), true)
 
 /**
- * Returns the function composition of the `fs` functions.
+ * Returns function composition of `f`.
+ *
+ * The result is `(...args: Parameters<FL>): ReturnType<FF> => f[0](f[1](...(f[n - 1](...args))...))`
  */
 export const compose =
 	<FF extends BasicFunction, FL extends AnyFunction>(
-		...fs: [FF, ...AnyFunction[], FL]
+		...f: [FF, ...AnyFunction[], FL]
 	) =>
 	(...x: Parameters<FL>): ReturnType<FF> => {
-		const start: ReturnType<FL> = ((last(fs) || id) as FL)(...x)
-		return lastOut(fs).reduceRight((last: any, curr: any) => curr(last), start)
+		const start: ReturnType<FL> = ((last(f) || id) as FL)(...x)
+		return lastOut(f).reduceRight((last: any, curr: any) => curr(last), start)
 	}
 
 /**
- * Returns the array, containing the values of `f(j * i)` for `0 <= i <= floor(n / j)`
+ * Returns the array, with values of `arr[i] = j * i` for `0 <= i <= floor(n / j)`.
+ *
+ * Default `j` value is `1`.
  */
-export const iterations = <T = any>(f: (i: number) => T, n: number, j = 1) =>
-	Array.from({ length: Math.floor(n / j) + +(j > 1) }, (_x, i) => f(j * i))
+export const jumps = (n: number, j = 1) =>
+	Array.from({ length: Math.floor(n / j) + +(j > 1) }, (_x, i) => j * i)
 
 /**
- * Creates a new array `X` = [init],
- * that is then filled with `n` iterations of `f(last(X), i, X)`,
- * and returned
+ * Returns an array `X`, with `X[0] = init` and `X[i] = f(X[i - 1], i - 1, X)`
+ * for `1 <= i <= n - 1`.
  */
 export const sequence =
 	<T = any>(f: (v: T, i: number, seq: T[]) => T, n: number) =>
 	(init: T) => {
-		const seqres = [init]
-		for (let i = 0; i < n; ++i) seqres.push(f(last(seqres), i, seqres))
-		return seqres
+		const seq = [init]
+		for (let i = 0; i < n; ++i) seq.push(f(seq[i], i, seq))
+		return seq
 	}
 
 /**
- * Makes the calls `f(0), f(1), ..., f(n - 1)`
+ * Makes the calls `f(x_0), f(x_1), ..., f(x_{n - 1})`, where `x_i` is the `i`th element of `x`
  */
-export const repeat = (f: (i: number) => void, n: number) => {
-	for (let i = 0; i < n; ++i) f(i)
+export const loopOver = <T = any>(f: (x: T) => void, x: Iterable<T>) => {
+	for (const xCurr of x) f(xCurr)
 }
 
 /**
- * Returns a composition, such that expected output of the given functions are arrays
- * intended to be spread out as inputs for the next function.
+ * Returns a composition, such that expected output of any of the given functions
+ * is an array intended to be spread out as input for the next function (i.e. one
+ * closer to the beginning of the array).
  */
 export const arrayCompose =
 	<FF extends AnyFunction, FL extends ArrayMapper>(
@@ -81,61 +88,53 @@ export const arrayCompose =
 	}
 
 /**
- * Creates and returns a `Map`, filled with key-value pairs of `[x, f(x)]` with `x` being in `keys`
+ * Returns a `Map`, filled with key-value pairs of `[x, f(x, i, keys)]` for each `x` in `keys`
  */
-export const cache = <K = any, F extends AnyFunction = AnyFunction>(
-	f: F,
-	keys: K[],
-): Map<K, ReturnType<F>> => new Map(keys.map((x) => [x, f(x)]))
+export const toMap = <Key = any, Out = any>(
+	f: (x: Key, i: number, keys: readonly Key[]) => Out,
+	keys: readonly Key[],
+): Map<Key, Out> => new Map(keys.map((x, i) => [x, f(x, i, keys)]))
 
 /**
- * Breaks the given inputs' array `x` into (possibly intersecting) segments using `x.slice(inds[i])`,
- * then - calls the respective function with each segment, and returns all of the calls'
+ * Breaks the given inputs' array `x` into (possibly intersecting) segments using `x.slice(ind[i])`,
+ * then - calls the respective function `f[i]` with each segment, and returns all of the calls'
  * return values in an array.
  *
  * A missing `x.slice`-interval defaults to `[]` (whole signature copying)
  */
 export const tupleSlice =
-	<FunctionTuple extends AnyFunction[]>(...fs: FunctionTuple) =>
-	(...inds: ([number?, number?] | undefined | null)[]) =>
-	(...x: any[]): MapReturnType<FunctionTuple> =>
-		fs.map((f, i) => f(...x.slice(...(inds[i] || [])))) as any
+	<FunctionTuple extends AnyFunction[]>(...f: FunctionTuple) =>
+	(...ind: ([number?, number?] | undefined | null)[]) =>
+	(...x: any[]) =>
+		f.map((fun, i) =>
+			fun(...x.slice(...(ind[i] || []))),
+		) as MapReturnType<FunctionTuple>
 
 /**
  * Similar to `tupleSlice`, only difference being that `inds` are now predicates,
  * using which the signatures for each particular function are `x.filter(inds[i])`-ed
- *
- * Like with `tupleSlice`, the default is copying of the entire signature `.filter(T)`
  */
 export const tuplePick =
 	<FunctionTuple extends AnyFunction[]>(...fs: FunctionTuple) =>
-	(
-		...inds: (
-			| ((value?: any, index?: number, array?: any[]) => any)
-			| null
-			| undefined
-		)[]
-	) =>
-	(...x: any[]) =>
-		fs.map((f, i) => f(...x.filter(inds[i] || T))) as Tuple<
-			any,
-			FunctionTuple["length"]
-		>
+	<In extends any[] = any[]>(...ind: Filter[]) =>
+	(...x: In) =>
+		fs.map((f, i) => f(...x.filter(ind[i]))) as MapReturnType<FunctionTuple>
 
 /**
- * Returns a function, the inputs of which are cached on each call.
+ * Returns a function inputs of which are cached on each call in a `Map<K, V>`.
  */
-export const cached = <K = any, V = any>(base: (x: K) => V) => {
-	const localCache = new Map<K, V>()
-	const cachef = function (x: K) {
-		const cacheval = cachef.cache.get(x)
+export const cached = <K = any, V = any>(base: (x: K) => NonNullable<V>) => {
+	const cache = new Map<K, V>()
+
+	function toCache(x: K) {
+		const cacheval = cache.get(x)
 		if (!isUndefined(cacheval)) return cacheval
-		const retres = base(x)
-		cachef.cache.set(x, retres)
-		return retres
+		const retval = base(x)
+		cache.set(x, retval)
+		return retval
 	}
-	cachef.cache = localCache
-	return cachef
+
+	return toCache
 }
 
 /**
@@ -149,38 +148,41 @@ export const id = <Type = any>(x: Type) => x
 export const nil = () => {}
 
 /**
- * Creates a new function calling `f`, with arguments at positions defined by `indexes` Set
- * filled with values from `values`, and the remaining arguments filled with values from 'x'.
+ * Returns a new function calling `f`, with arguments at positions defined by `indexes` Set
+ * filled with `start`, and the remaining arguments filled with 'rest'.
  */
 export const argFiller = (f: AnyFunction) => {
 	return (...indexes: number[]) => {
 		const substitutionForm = substitute(f.length, indexes)
-		return (...values: any[]) => {
-			const substituter = substitutionForm(values)
-			return (...x: any[]) => f(...substituter(x))
+		return (...start: any[]) => {
+			const restPlugger = substitutionForm(start)
+			return (...rest: any[]) => f(...restPlugger(rest))
 		}
 	}
 }
 
 /**
- * Returns `f.bind(bound)`. `bound` defaults to `null`
+ * Returns `f.bind(bound)`
  */
-export const copy = <F extends AnyFunction>(f: F, bound: any = null): F =>
+export const bind = <F extends AnyFunction>(f: F, bound: any): F =>
 	f.bind(bound)
 
 /**
  * Returns a function returning `set.has(x)`
  */
-export const has = (set: Set<any>) => (x: any) => set.has(x)
+export const has =
+	<T = any>(set: Set<T>) =>
+	(x: T) =>
+		set.has(x)
 
 /**
- * Returns a function that removes last `max(n, args.length)` arguments
+ * Returns a function removing the last `min(n, args.length)` arguments
  * and calls `f` on the result. Default `n` is `Infinity`.
  */
 export const argWaster =
 	(n: number = Infinity) =>
-	(f: AnyFunction) =>
-	(...args: any[]) =>
+	<F extends AnyFunction = AnyFunction>(f: F) =>
+	(...args: any[]): ReturnType<F> =>
 		f(...args.slice(0, max(0, args.length - n)))
 
 /**